requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.0
cryptography==2.5
docopt==0.6.2
pycparser==2.19
PyNaCl==1.3.0
PyYAML==3.13
six==1.12.0
